refactor program model imports to use new location
