+++ registrar/apps/api/serializers.py
@@ -7,11 +7,11 @@ in question should be moved to versioned sub-package.
 from rest_framework import serializers
 from user_tasks.models import UserTaskStatus
 
+from registrar.apps.core.models import Program
 from registrar.apps.enrollments.constants import (
     COURSE_ENROLLMENT_STATUSES,
     PROGRAM_ENROLLMENT_STATUSES,
 )
-from registrar.apps.enrollments.models import Program
 
 
 # pylint: disable=abstract-method
