+++ registrar/apps/common/tasks.py
@@ -9,7 +9,7 @@ from user_tasks.tasks import UserTask
 from registrar.apps.core.constants import UPLOADS_PATH_PREFIX
 from registrar.apps.core.filestore import get_filestore
 from registrar.apps.core.jobs import post_job_failure
-from registrar.apps.enrollments.models import Program
+from registrar.apps.core.models import Program
 
 
 log = get_task_logger(__name__)
