--- registrar/apps/enrollments/models.py
@@ -1,4 +1,0 @@
-"""
-Temporary proxy for importing the Program model
-"""
-from registrar.apps.core.models import Program  # pylint: disable=unused-import
